@@ -23,7 +23,12 @@ dependencies:
   carousel_pro: ^1.0.0
   flutter_swiper: ^1.1.6
   flutter_screenutil: ^0.5.3
-  google_maps_flutter: 0.2.0
+  google_maps_flutter: ^0.5.11
+  geolocator: ^5.1.5
+  font_awesome_flutter: ^8.2.0
+
+
+
 
 
 
