@@ -7,6 +7,8 @@ class PublicSite {
   String description;
   String brand;
   String title;
+  double lat;
+  double lon;
 
 
 
@@ -18,6 +20,8 @@ class PublicSite {
     this.description,
     this.brand,
     this.title,
+    this.lat,
+    this.lon,
   });
 }
 
@@ -27,37 +31,52 @@ class PublicSite {
  List<PublicSite> iglesiasSitios = [
 
     PublicSite(
-      title: 'Iglesias',
+      category: 'Iglesias',
       IdSite: 1,
       brand: "Catolica",
-      name: "Iglesia del Carmen ",
-      image: 'assets/iglesias/carmen.jpg',
+      name: 'Iglesia del Carmen ',
+      lat: 8.981511,
+      lon: -79.528074,
+      image: "assets/iglesias/carmen.jpg",
       description:
           "Es icono arquitectónico de Bella Vista. La Iglesia Nuestra Señora del Carmen engalana la vía España con su imponente estructura de estilo gótico medieval desde hace más de 62 años.",
     ),
     PublicSite(
-      title: 'Iglesias',
+      category: 'Iglesias',
       IdSite: 2,
       brand: "Catolica",
       name: "San Francisco de Asís",
+      lat: 8.953283,
+      lon: -79.533261,
       image: 'assets/iglesias/asis.jpg',
       description: "construcion",
     ),
     PublicSite(
-      title: 'Iglesias',
+      category: 'Iglesias',
       IdSite: 3,
       brand: "Musulmana",
       name: "Jumma Masjid",
+      lat: 8.966633,
+      lon: -79.535344,
       image: 'assets/iglesias/jumma.jpg',
       description:
           "Es icono arquitectónico de Bella Vista. La Iglesia Nuestra Señora del Carmen engalana la vía España con su imponente estructura de estilo gótico medieval desde hace más de 62 años.",
     ),
+   PublicSite(
+     category: 'Parques',
+     IdSite: 4,
+     brand: "Entretenimiento",
+     name: "Parque urraca ",
+     image: 'assets/parques/urraca.jpg',
+     description:
+     "Es icono arquitectónico de Bella Vista. La Iglesia Nuestra Señora del Carmen engalana la vía España con su imponente estructura de estilo gótico medieval desde hace más de 62 años.",
+   ),
   ];
 
 List<PublicSite> parques = [
 
   PublicSite(
-    title: 'Parques',
+    title: 'Parques2',
     IdSite: 4,
     brand: "Entretenimiento",
     name: "Parque urraca ",
